Add test helper for ROS stubs
